Refactor CloudTrail
